IRGen: refactor APInt bit pattern chunking
This change removes most of the code that 'chunks' APInts by
masking out the low N bits and then shifting the value right by
N bits in a loop. The shifting is replaced by a new APIntReader
class that allows the APInt to be 'chunked' in high-to-low order
(i.e. big-endian) as well as the existing low-to-high (i.e.
little-endian) order.